3.0.0: make async with optional sync
