+++ .travis.yml
@@ -1,13 +1,11 @@
 sudo: false
 language: node_js
 node_js:
-  - "0.12"
-  - "4"
-  - "5"
-  - "6"
+  - "8"
+  - "lts/*"
 matrix:
   include:
-    - node_js: "6"
+    - node_js: "8"
       env: TEST_SUITE=lint
 env:
   - TEST_SUITE=unit
