+++ README.md
@@ -26,13 +26,20 @@ Example
 -------
 
 ```js
-var scrypt = require('scryptsy')
+const { scrypt, scryptSync } = require('scryptsy')
 
-var key = "pleaseletmein"
-var salt = "SodiumChloride"
-var data = scrypt(key, salt, 16384, 8, 1, 64)
-console.log(data.toString('hex'))
-// => 7023bdcb3afd7348461c06cd81fd38ebfda8fbba904f8e3ea9b543f6545da1f2d5432955613f0fcf62d49705242a9af9e61e85dc0d651e40dfcf017b45575887
+async function main () {
+  var key = "pleaseletmein"
+  var salt = "SodiumChloride"
+  var data1 = scryptSync(key, salt, 16384, 8, 1, 64)
+  console.log(data1.toString('hex'))
+  // => 7023bdcb3afd7348461c06cd81fd38ebfda8fbba904f8e3ea9b543f6545da1f2d5432955613f0fcf62d49705242a9af9e61e85dc0d651e40dfcf017b45575887
+
+  var data2 = await scrypt(key, salt, 16384, 8, 1, 64)
+  console.log(data2.toString('hex'))
+  // => 7023bdcb3afd7348461c06cd81fd38ebfda8fbba904f8e3ea9b543f6545da1f2d5432955613f0fcf62d49705242a9af9e61e85dc0d651e40dfcf017b45575887
+}
+main().catch(console.error)
 ```
 
 
@@ -40,6 +47,18 @@ API
 ---
 
 ### scrypt(key, salt, N, r, p, keyLenBytes, [progressCallback])
+
+- **key**: The key. Either `Buffer` or `string`.
+- **salt**: The salt. Either `Buffer` or `string`.
+- **N**: The number of iterations. `number` (integer)
+- **r**: Memory factor. `number` (integer)
+- **p**: Parallelization factor. `number` (integer)
+- **keyLenBytes**: The number of bytes to return. `number` (integer)
+- **progressCallback**: Call callback on every `1000` ops. Passes in `{current, total, percent}` as first parameter to `progressCallback()`.
+
+Returns `Promise<Buffer>`.
+
+### scryptSync(key, salt, N, r, p, keyLenBytes, [progressCallback])
 
 - **key**: The key. Either `Buffer` or `string`.
 - **salt**: The salt. Either `Buffer` or `string`.
