+++ lib/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  scrypt: require('./scrypt'),
+  scryptSync: require('./scryptSync')
+}
