+++ package.json
@@ -1,16 +1,16 @@
 {
   "name": "scryptsy",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "description": "Pure JavaScript implementation of the scrypt key deriviation function that is fully compatible with Node.js and the browser.",
-  "main": "lib/scrypt.js",
+  "main": "lib/index.js",
   "author": "",
   "license": "MIT",
   "devDependencies": {
-    "coveralls": "^2.10.0",
-    "istanbul": "^0.3.5",
-    "mocha": "^2.2.0",
-    "mochify": "^2.1.0",
-    "standard": "^7.1.1"
+    "coveralls": "^3.0.3",
+    "mocha": "^6.0.2",
+    "mochify": "^6.1.0",
+    "nyc": "^13.3.0",
+    "standard": "^12.0.1"
   },
   "dependencies": {},
   "repository": {
@@ -29,10 +29,10 @@
   ],
   "scripts": {
     "test": "mocha --ui bdd",
-    "unit": "./node_modules/.bin/mocha",
-    "coverage": "./node_modules/.bin/istanbul cover ./node_modules/.bin/_mocha -- --reporter list test/*.js",
-    "coveralls": "npm run-script coverage && node ./node_modules/.bin/coveralls < coverage/lcov.info",
-    "browser-test": "./node_modules/.bin/mochify --wd -R spec",
+    "unit": "mocha",
+    "coverage": "nyc --check-coverage --statements 80 --branches 60 --functions 90 --lines 90 mocha",
+    "coveralls": "npm run-script coverage && coveralls < coverage/lcov.info",
+    "browser-test": "mochify --wd -R spec",
     "lint": "standard"
   }
 }
